requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg[binary]==3.2.12
feedparser==6.0.12
chromadb==1.0.20
openai==1.107.0
pydantic-settings==2.11.0
python-dateutil==2.9.0.post0
pytest==8.4.2
